move test into selenium lair
