+++ tests/modules/selenium/test_selenium_builder.py
@@ -1,11 +1,19 @@
 import os
-
+import yaml
 from bzt.six import PY2
 from tests import RESOURCES_DIR
 from tests.modules.selenium import SeleniumTestCase
 
 
 class TestSeleniumScriptGeneration(SeleniumTestCase):
+
+    def test_new_flow_build(self):
+        config = yaml.load(open('wip/new-flow/a.yml').read())
+        self.configure(config)
+        self.obj.prepare()
+        exp_file = RESOURCES_DIR + "selenium/test_new_flow.py"
+        self.assertFilesEqual(exp_file, self.obj.script, replace_str=self.obj.wdlog, replace_with="webdriver.log", python_files=True)
+
     def test_modern_actions_generator(self):
         self.configure({
             "execution": [{
